refactor(synthesis): deduplicate limit constants and consolidate tests
- Export MAX_SYNTH_ARTIFACT_ITEMS and MAX_CONTEXT_EXPANSION_ITEMS from
  lib/listening-sessions/synthesis.ts as single source of truth
- Import shared constants in convex/listeningSessions.ts instead of
  maintaining independent copies (fixes the exact duplication that
  caused the original 8-vs-6 mismatch bug)
- Merge unique test from synthesis-partial-failure.test.ts into
  synthesis.test.ts (partial object normalization)
- Delete synthesis-partial-failure.test.ts (3 of 5 tests were
  redundant with existing synthesis.test.ts coverage)
